requirements.txt
# --- Core app/UI ---
streamlit==1.37.0
tqdm

# --- Audio / ASR ---
faster-whisper==1.0.3
ffmpeg-python
imageio-ffmpeg
pydub
soundfile

# --- ML stack (CPU-friendly pins for Py3.11) ---
torch==2.3.1
torchaudio==2.3.1
numpy==1.26.4

# --- Embeddings / retrieval ---
sentence-transformers==3.0.1
chromadb==0.5.5
pypika==0.48.9

# --- Evaluation / datasets ---
ragas==0.1.14
pyarrow==14.0.2

# --- Optional diarization (only used if HF_TOKEN is set) ---
pyannote.audio==3.1.1
julius==0.2.7
omegaconf==2.3.0
antlr4-python3-runtime==4.9.3
